refactor: move LoginPage to proper dir
