+++ src/App.tsx
@@ -17,7 +17,6 @@ import {
 } from 'react-router-dom'
 import { toast } from 'sonner'
 import { useStore } from 'zustand'
-import LoginPage from '@/components/LoginPage'
 import { LogsPage } from '@/components/LogsPage'
 import MainWalletPage from '@/components/MainWalletPage'
 import SwitchWalletPage from '@/components/SwitchWalletPage'
@@ -25,6 +24,7 @@ import CreateWalletPage from '@/components/create/CreateWalletPage'
 import { EarnPage } from '@/components/earn/EarnPage'
 import ErrorPage from '@/components/error/ErrorPage'
 import { Layout } from '@/components/layout/Layout'
+import LoginPage from '@/components/login/LoginPage'
 import { OrderbookPage } from '@/components/orderbook/OrderbookPage'
 import { ReceivePage } from '@/components/receive/ReceivePage'
 import { SendPage } from '@/components/send/SendPage'
